backend/requirements.txt
# requirements.txt for backend
fastapi
uvicorn
pandas
scikit-learn
xgboost
python-dotenv
joblib
numpy